requirements.txt
aiosqlite
arrow
discord.py
jishaku
python-dotenv
uvloop ; sys_platform == "linux"